v0.7311117

+++ modules/mark/modules/strategys.py
@@ -28,6 +28,19 @@ def parseArgs(args):
     return parser.parse_args(args)
 
 
+def print_range(train_bars, bband_list, ave_list, data_range, split):
+    totallenth = len(train_bars.symbol_ori_data[train_bars.symbol_list[0]]["close"])
+    mid_lenth = max(bband_list[-1] - 1, ave_list[-1])
+    trainpre_pos = max(data_range[0] if data_range[0] is not None else 0, ave_list[-1] - 1)
+    validaft_lenth = max(data_range[1] if data_range[1] is not None else 0, bband_list[-1])
+    usefull_lenth = totallenth - trainpre_pos - validaft_lenth - mid_lenth
+    trainaft_pos = int(usefull_lenth * split) + trainpre_pos
+    validpre_pos = trainpre_pos + int(usefull_lenth * split) + mid_lenth
+    validaft_pos = trainpre_pos + usefull_lenth + mid_lenth
+    print("total length: {} train range:{}-{}. valid range:{}-{}.".format(totallenth, trainpre_pos, trainaft_pos,
+                                                                          validpre_pos, validaft_pos))
+
+
 # 封装对数据的计算，并且生成相应的信号:  策略处理基类，可用于处理历史和实际交易数据，只需把数据存到队列中。
 # 移动平均跨越策略。用短期/长期移动平均值进行基本的移动平均跨越的实现。
 class MovingAverageCrossStrategy(strategy.BacktestingStrategy):
@@ -768,7 +781,8 @@ class MlaStrategy(strategy.BacktestingStrategy):
         globalstep = modelcrnn.batch_train(inputs_t, targets_base_t, targets_much_t, inputs_v, targets_base_v,
                                            targets_much_v, batch_size, num_epochs)
 
-    def train_probability_everysignals(self, train_bars, ave_list, bband_list, date_range, split=0.8, args=None):
+    def train_probability_everysignals(self, train_bars, ave_list, bband_list, date_range, newdata=1, split=0.8,
+                                       args=None):
         """训练"""
         # 1. 输入参数
         self._prepare_model_para(args)
@@ -776,25 +790,7 @@ class MlaStrategy(strategy.BacktestingStrategy):
         data_buff_dir = "everynpy_" + "_".join([str(i1) for i1 in bband_list])
         full_data_buff_dir = os.path.join(data_path, data_buff_dir)
         makesurepath(full_data_buff_dir)
-        if os.path.isfile(os.path.join(full_data_buff_dir, "drawdw_v.npy")):
-            print("loadingdata")
-            inputs_t = np.load(os.path.join(full_data_buff_dir, "inputs_t.npy"))
-            reta_t = np.load(os.path.join(full_data_buff_dir, "reta_t.npy"))
-            reth_t = np.load(os.path.join(full_data_buff_dir, "reth_t.npy"))
-            retl_t = np.load(os.path.join(full_data_buff_dir, "retl_t.npy"))
-            stdup_t = np.load(os.path.join(full_data_buff_dir, "stdup_t.npy"))
-            stddw_t = np.load(os.path.join(full_data_buff_dir, "stddw_t.npy"))
-            drawup_t = np.load(os.path.join(full_data_buff_dir, "drawup_t.npy"))
-            drawdw_t = np.load(os.path.join(full_data_buff_dir, "drawdw_t.npy"))
-            inputs_v = np.load(os.path.join(full_data_buff_dir, "inputs_v.npy"))
-            reta_v = np.load(os.path.join(full_data_buff_dir, "reta_v.npy"))
-            reth_v = np.load(os.path.join(full_data_buff_dir, "reth_v.npy"))
-            retl_v = np.load(os.path.join(full_data_buff_dir, "retl_v.npy"))
-            stdup_v = np.load(os.path.join(full_data_buff_dir, "stdup_v.npy"))
-            stddw_v = np.load(os.path.join(full_data_buff_dir, "stddw_v.npy"))
-            drawup_v = np.load(os.path.join(full_data_buff_dir, "drawup_v.npy"))
-            drawdw_v = np.load(os.path.join(full_data_buff_dir, "drawdw_v.npy"))
-        else:
+        if 1 == newdata:
             # 2. 加载衍生前值
             train_bars.generate_b_derivative()
             # 2. 加载衍生后值
@@ -802,22 +798,51 @@ class MlaStrategy(strategy.BacktestingStrategy):
             inputs_t, reta_t, reth_t, retl_t, stdup_t, stddw_t, drawup_t, drawdw_t, inputs_v, reta_v, reth_v, retl_v, \
             stdup_v, stddw_v, drawup_v, drawdw_v = self._prepare_every_train_data(train_bars, ave_list, bband_list,
                                                                                   date_range, split)
-            np.save(os.path.join(full_data_buff_dir, "inputs_t"), inputs_t)
-            np.save(os.path.join(full_data_buff_dir, "reta_t"), reta_t)
-            np.save(os.path.join(full_data_buff_dir, "reth_t"), reth_t)
-            np.save(os.path.join(full_data_buff_dir, "retl_t"), retl_t)
-            np.save(os.path.join(full_data_buff_dir, "stdup_t"), stdup_t)
-            np.save(os.path.join(full_data_buff_dir, "stddw_t"), stddw_t)
-            np.save(os.path.join(full_data_buff_dir, "drawup_t"), drawup_t)
-            np.save(os.path.join(full_data_buff_dir, "drawdw_t"), drawdw_t)
-            np.save(os.path.join(full_data_buff_dir, "inputs_v"), inputs_v)
-            np.save(os.path.join(full_data_buff_dir, "reta_v"), reta_v)
-            np.save(os.path.join(full_data_buff_dir, "reth_v"), reth_v)
-            np.save(os.path.join(full_data_buff_dir, "retl_v"), retl_v)
-            np.save(os.path.join(full_data_buff_dir, "stdup_v"), stdup_v)
-            np.save(os.path.join(full_data_buff_dir, "stddw_v"), stddw_v)
-            np.save(os.path.join(full_data_buff_dir, "drawup_v"), drawup_v)
-            np.save(os.path.join(full_data_buff_dir, "drawdw_v"), drawdw_v)
+        else:
+            if os.path.isfile(os.path.join(full_data_buff_dir, "drawdw_v.npy")):
+                print("loadingdata")
+                inputs_t = np.load(os.path.join(full_data_buff_dir, "inputs_t.npy"))
+                reta_t = np.load(os.path.join(full_data_buff_dir, "reta_t.npy"))
+                reth_t = np.load(os.path.join(full_data_buff_dir, "reth_t.npy"))
+                retl_t = np.load(os.path.join(full_data_buff_dir, "retl_t.npy"))
+                stdup_t = np.load(os.path.join(full_data_buff_dir, "stdup_t.npy"))
+                stddw_t = np.load(os.path.join(full_data_buff_dir, "stddw_t.npy"))
+                drawup_t = np.load(os.path.join(full_data_buff_dir, "drawup_t.npy"))
+                drawdw_t = np.load(os.path.join(full_data_buff_dir, "drawdw_t.npy"))
+                inputs_v = np.load(os.path.join(full_data_buff_dir, "inputs_v.npy"))
+                reta_v = np.load(os.path.join(full_data_buff_dir, "reta_v.npy"))
+                reth_v = np.load(os.path.join(full_data_buff_dir, "reth_v.npy"))
+                retl_v = np.load(os.path.join(full_data_buff_dir, "retl_v.npy"))
+                stdup_v = np.load(os.path.join(full_data_buff_dir, "stdup_v.npy"))
+                stddw_v = np.load(os.path.join(full_data_buff_dir, "stddw_v.npy"))
+                drawup_v = np.load(os.path.join(full_data_buff_dir, "drawup_v.npy"))
+                drawdw_v = np.load(os.path.join(full_data_buff_dir, "drawdw_v.npy"))
+                # 打印尺寸
+                print_range(train_bars, bband_list, ave_list, date_range, split)
+            else:
+                # 2. 加载衍生前值
+                train_bars.generate_b_derivative()
+                # 2. 加载衍生后值
+                train_bars.generate_a_derivative()
+                inputs_t, reta_t, reth_t, retl_t, stdup_t, stddw_t, drawup_t, drawdw_t, inputs_v, reta_v, reth_v, retl_v, \
+                stdup_v, stddw_v, drawup_v, drawdw_v = self._prepare_every_train_data(train_bars, ave_list, bband_list,
+                                                                                      date_range, split)
+                np.save(os.path.join(full_data_buff_dir, "inputs_t"), inputs_t)
+                np.save(os.path.join(full_data_buff_dir, "reta_t"), reta_t)
+                np.save(os.path.join(full_data_buff_dir, "reth_t"), reth_t)
+                np.save(os.path.join(full_data_buff_dir, "retl_t"), retl_t)
+                np.save(os.path.join(full_data_buff_dir, "stdup_t"), stdup_t)
+                np.save(os.path.join(full_data_buff_dir, "stddw_t"), stddw_t)
+                np.save(os.path.join(full_data_buff_dir, "drawup_t"), drawup_t)
+                np.save(os.path.join(full_data_buff_dir, "drawdw_t"), drawdw_t)
+                np.save(os.path.join(full_data_buff_dir, "inputs_v"), inputs_v)
+                np.save(os.path.join(full_data_buff_dir, "reta_v"), reta_v)
+                np.save(os.path.join(full_data_buff_dir, "reth_v"), reth_v)
+                np.save(os.path.join(full_data_buff_dir, "retl_v"), retl_v)
+                np.save(os.path.join(full_data_buff_dir, "stdup_v"), stdup_v)
+                np.save(os.path.join(full_data_buff_dir, "stddw_v"), stddw_v)
+                np.save(os.path.join(full_data_buff_dir, "drawup_v"), drawup_v)
+                np.save(os.path.join(full_data_buff_dir, "drawdw_v"), drawdw_v)
         # 3. 训练
         print(inputs_t.shape, reta_t.shape, reth_t.shape, retl_t.shape,
               stdup_t.shape, stddw_t.shape, drawup_t.shape, drawdw_t.shape,
